requirements.txt
PyMuPDF
Pillow
reportlab
